refactor: use MessageSquarePlus icon, rename hook file to useSummarizeInNewChat.ts
